Move porous flow unit tests to porous_flow/unit
Closes #14815
