+++ modules/porous_flow/unit/include/PorousFlowFluidStateFlashTest.h
@@ -11,21 +11,21 @@
 
 #include "MooseObjectUnitTest.h"
 #include "PorousFlowCapillaryPressureVG.h"
-#include "PorousFlowWaterVapor.h"
+#include "PorousFlowWaterNCG.h"
 #include "Water97FluidProperties.h"
+#include "CO2FluidProperties.h"
 
-class PorousFlowWaterVaporTest : public MooseObjectUnitTest
+class PorousFlowFluidStateFlashTest : public MooseObjectUnitTest
 {
 public:
-  PorousFlowWaterVaporTest() : MooseObjectUnitTest("MooseUnitApp") { buildObjects(); }
+  PorousFlowFluidStateFlashTest() : MooseObjectUnitTest("PorousFlowApp") { buildObjects(); }
 
 protected:
   void buildObjects()
   {
     InputParameters pc_params = _factory.getValidParams("PorousFlowCapillaryPressureVG");
     pc_params.set<Real>("m") = 0.5;
-    pc_params.set<Real>("alpha") = 1.0e-4;
-    pc_params.set<Real>("pc_max") = 1.0e5;
+    pc_params.set<Real>("alpha") = 0.1;
     _fe_problem->addUserObject("PorousFlowCapillaryPressureVG", "pc", pc_params);
     _pc = &_fe_problem->getUserObjectTempl<PorousFlowCapillaryPressureVG>("pc");
 
@@ -33,14 +33,20 @@ protected:
     _fe_problem->addUserObject("Water97FluidProperties", "water_fp", water_params);
     _water_fp = &_fe_problem->getUserObjectTempl<Water97FluidProperties>("water_fp");
 
-    InputParameters uo_params = _factory.getValidParams("PorousFlowWaterVapor");
+    InputParameters ncg_params = _factory.getValidParams("CO2FluidProperties");
+    _fe_problem->addUserObject("CO2FluidProperties", "ncg_fp", ncg_params);
+    _ncg_fp = &_fe_problem->getUserObjectTempl<CO2FluidProperties>("ncg_fp");
+
+    InputParameters uo_params = _factory.getValidParams("PorousFlowWaterNCG");
     uo_params.set<UserObjectName>("water_fp") = "water_fp";
+    uo_params.set<UserObjectName>("gas_fp") = "ncg_fp";
     uo_params.set<UserObjectName>("capillary_pressure") = "pc";
-    _fe_problem->addUserObject("PorousFlowWaterVapor", "fp", uo_params);
-    _fp = &_fe_problem->getUserObjectTempl<PorousFlowWaterVapor>("fp");
+    _fe_problem->addUserObject("PorousFlowWaterNCG", "fp", uo_params);
+    _fp = &_fe_problem->getUserObjectTempl<PorousFlowWaterNCG>("fp");
   }
 
   const PorousFlowCapillaryPressureVG * _pc;
-  const PorousFlowWaterVapor * _fp;
+  const PorousFlowWaterNCG * _fp;
   const Water97FluidProperties * _water_fp;
+  const CO2FluidProperties * _ncg_fp;
 };
