+++ modules/porous_flow/unit/include/PorousFlowWaterVaporTest.h
@@ -11,14 +11,13 @@
 
 #include "MooseObjectUnitTest.h"
 #include "PorousFlowCapillaryPressureVG.h"
-#include "PorousFlowWaterNCG.h"
+#include "PorousFlowWaterVapor.h"
 #include "Water97FluidProperties.h"
-#include "CO2FluidProperties.h"
 
-class PorousFlowWaterNCGTest : public MooseObjectUnitTest
+class PorousFlowWaterVaporTest : public MooseObjectUnitTest
 {
 public:
-  PorousFlowWaterNCGTest() : MooseObjectUnitTest("MooseUnitApp") { buildObjects(); }
+  PorousFlowWaterVaporTest() : MooseObjectUnitTest("PorousFlowApp") { buildObjects(); }
 
 protected:
   void buildObjects()
@@ -26,35 +25,22 @@ protected:
     InputParameters pc_params = _factory.getValidParams("PorousFlowCapillaryPressureVG");
     pc_params.set<Real>("m") = 0.5;
     pc_params.set<Real>("alpha") = 1.0e-4;
+    pc_params.set<Real>("pc_max") = 1.0e5;
     _fe_problem->addUserObject("PorousFlowCapillaryPressureVG", "pc", pc_params);
     _pc = &_fe_problem->getUserObjectTempl<PorousFlowCapillaryPressureVG>("pc");
 
     InputParameters water_params = _factory.getValidParams("Water97FluidProperties");
     _fe_problem->addUserObject("Water97FluidProperties", "water_fp", water_params);
-    _water_fp = &_fe_problem->getUserObjectTempl<SinglePhaseFluidProperties>("water_fp");
+    _water_fp = &_fe_problem->getUserObjectTempl<Water97FluidProperties>("water_fp");
 
-    InputParameters ncg_params = _factory.getValidParams("CO2FluidProperties");
-    _fe_problem->addUserObject("CO2FluidProperties", "ncg_fp", ncg_params);
-    _ncg_fp = &_fe_problem->getUserObjectTempl<SinglePhaseFluidProperties>("ncg_fp");
-
-    InputParameters uo_params = _factory.getValidParams("PorousFlowWaterNCG");
+    InputParameters uo_params = _factory.getValidParams("PorousFlowWaterVapor");
     uo_params.set<UserObjectName>("water_fp") = "water_fp";
-    uo_params.set<UserObjectName>("gas_fp") = "ncg_fp";
     uo_params.set<UserObjectName>("capillary_pressure") = "pc";
-    _fe_problem->addUserObject("PorousFlowWaterNCG", "fs", uo_params);
-    _fs = &_fe_problem->getUserObjectTempl<PorousFlowWaterNCG>("fs");
-
-    // Indices for derivatives
-    _pidx = _fs->getPressureIndex();
-    _Tidx = _fs->getTemperatureIndex();
-    _Zidx = _fs->getZIndex();
+    _fe_problem->addUserObject("PorousFlowWaterVapor", "fp", uo_params);
+    _fp = &_fe_problem->getUserObjectTempl<PorousFlowWaterVapor>("fp");
   }
 
   const PorousFlowCapillaryPressureVG * _pc;
-  const PorousFlowWaterNCG * _fs;
-  const SinglePhaseFluidProperties * _water_fp;
-  const SinglePhaseFluidProperties * _ncg_fp;
-  unsigned int _pidx;
-  unsigned int _Tidx;
-  unsigned int _Zidx;
+  const PorousFlowWaterVapor * _fp;
+  const Water97FluidProperties * _water_fp;
 };
